Closed #22 - Add common functions - Also eitherT() has been removed
