+++ src/main/scala/kevinlee/fp/Fp.scala
@@ -1,0 +1,7 @@
+package kevinlee.fp
+
+/**
+  * @author Kevin Lee
+  * @since 2019-06-02
+  */
+object Fp extends EitherFunctions
